rag/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-ai-inference==1.0.0b9
azure-core==1.32.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dotenv==0.9.9
fastapi==0.104.1
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.2
openai==1.68.2
packaging==24.2
pillow==11.1.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
pydantic==2.5.2
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentence-transformers==2.2.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.24.0
# uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
aiohttp==3.9.1
faiss-cpu==1.7.4
llama-cpp-python==0.2.23